comprimido.py
# convert_multi_format_v6.py
# Conversor modular multi-formato para Docusaurus — v6.1 (SLIM & IMPROVED)
# Soporta: PDF, DOCX, XLSX, XLS, XLSM, PPTX, TXT, PNG, ZIP
#
# NOVEDADES v6.1:
#   - 🛡️ TABLAS PROTEGIDAS: Placeholder mechanism mejorado.
#   - 🧹 FIRMAS LIMPIAS: Eliminación precisa de certificados y fechas de firma.
#   - 🔤 NOMBRES CONTEXTUALES: Normalización que preserva términos clave para SEO.
#   - ⚖️ JUSTIFICACIÓN PRO: Regex optimizada para párrafos de texto continuo.
#   - ⚡ CÓDIGO SLIM: Estructura refinada para mayor legibilidad y mantenimiento.

import re, time, unicodedata, logging, tempfile, shutil, warnings
from pathlib import Path
from concurrent.futures import ThreadPoolExecutor, as_completed
from typing import Optional, Tuple, List, Dict
from datetime import datetime

# Lazy loading of heavy dependencies
try:
    import fitz, pymupdf4llm
    PYMUPDF_OK = True
except: PYMUPDF_OK = False

try:
    from docx import Document
    DOCX_OK = True
except: DOCX_OK = False

try:
    import pandas as pd
    PANDAS_OK = True
except: PANDAS_OK = False

try:
    from pptx import Presentation
    PPTX_OK = True
except: PPTX_OK = False

try:
    from PIL import Image
    import pytesseract
    OCR_OK = True
except: OCR_OK = False

# Global state
_MODES = {"debug": False, "ocr": False, "strict": False}
_stats = None
_logger = None

def debug_log(msg: str):
    if _MODES["debug"]: print(f"[DEBUG] {msg}")

class Stats:
    def __init__(self):
        self.data = {"files": {}, "chars": [0, 0], "tables": [0, 0], "signatures": 0, "ocr": 0, "time": 0.0}
    def show(self):
        if not _MODES["debug"]: return
        print("\n" + "─"*50 + "\n📊 MÉTRICAS v6.1\n" + "─"*50)
        for k, v in self.data["files"].items(): print(f" {k:8}: {v} docs")
        raw, clean = self.data["chars"]
        if raw: print(f" 📝 Chars: {raw:,} -> {clean:,} (-{(raw-clean)/raw*100:.1f}%)")
        if self.data["signatures"]: print(f" ✍️ Signatures: {self.data['signatures']}")
        if self.data["ocr"]: print(f" 🔍 OCR files: {self.data['ocr']}")
        print(f" ⏱ Time: {self.data['time']:.2f}s\n" + "─"*50)

# Utils
def normalize_fn(name: str) -> str:
    path = Path(name)
    stem = unicodedata.normalize("NFKD", path.stem).encode("ascii", "ignore").decode("ascii")
    # Conservar contexto (espacios/guiones -> -) y limpiar basura
    clean = re.sub(r"[^\w\s\-]", "", stem).strip()
    return re.sub(r"[\s_]+", "-", clean.lower()) + ".md"

def get_fm(p: Path) -> str:
    sid = re.sub(r"[^\w\-]", "-", p.stem.lower()).strip("-")
    title = p.stem.replace("-", " ").replace("_", " ").title()
    return f"---\nid: {sid}\ntitle: \"{title}\"\nsidebar_label: \"{title}\"\n---\n\n"

# Table Protection
def toggle_tables(text: str, tables: List[str] = None) -> Tuple[str, List[str]]:
    if tables is None: # Isolate
        extracted = []
        def repl(m):
            extracted.append(m.group(0))
            return f"\n<!-- TABLE_{len(extracted)-1} -->\n"
        # Detect table blocks
        res = re.sub(r"(?:^\|.+\n)+", repl, text, flags=re.M)
        return res, extracted
    # Restore
    for i, t in enumerate(tables):
        text = text.replace(f"<!-- TABLE_{i} -->", f"\n{t}\n")
    return text, []

# Cleaning Pipeline
def clean_md(text: str, verbose: bool = True) -> str:
    _stats.data["chars"][0] += len(text)
    
    # 1. Signatures
    orig = len(text)
    text = re.sub(r'`Firmado por[^`]+`', '[FIRMA DIGITAL]', text, flags=re.I)
    text = re.sub(r'\*{2,}\d+\*{2,}', '', text)
    text = re.sub(r'el día \d{2}/\d{2}/\d{4}.*$', '', text, flags=re.M)
    if len(text) < orig: _stats.data["signatures"] += 1
    
    # 2. Base MDX
    text = re.sub(r"<(?!/?(?:p|br|hr|img|div|span|strong|em|b|i|u|h\d)\b)", "&lt;", text)
    text = text.replace("{", "&#123;").replace("}", "&#125;")
    for tag in ["br", "hr"]: text = re.sub(fr"<{tag}\s*>", f"<{tag} />", text, flags=re.I)
    
    # 3. Tables
    text, tables = toggle_tables(text)
    
    # 4. Cleaning passthrough
    text = re.sub(r"([a-z])-\n([a-z])", r"\1\2", text, flags=re.I) # Hyphens
    text = re.sub(r"\s*\.{3,}\s*\d+", "", text) # Index dots
    
    if not _MODES["strict"]:
        # Paragraph merging
        lines = text.split("\n")
        merged = []
        i = 0
        while i < len(lines):
            l = lines[i].rstrip()
            if not l or l.startswith(("#","-","*","|",">","`","<")) or l.endswith((".",":",";","!","?")):
                merged.append(l)
                i += 1
            elif i+1 < len(lines) and lines[i+1].strip() and not lines[i+1].strip().startswith(("#","-","*","|",">","`","<")):
                merged.append(l + " " + lines[i+1].lstrip())
                i += 2
            else:
                merged.append(l)
                i += 1
        text = "\n".join(merged)

    # 5. Justify
    pattern_structural = r"^(?:#|\-|\*|\d+\.|\||```|<|id:|title:|sidebar_label:|---)"
    def just_cb(m):
        b = m.group(0).strip()
        if any(re.match(pattern_structural, l.strip()) for l in b.split('\n')) or len(b) < 60: return m.group(0)
        return f'\n<div style={{{{textAlign: "justify"}}}}>\n\n{b}\n\n</div>\n'
    text = re.sub(r"((?:(?!\n\n).)+)", just_cb, text, flags=re.S)
    
    # 6. Restore Tables
    text, _ = toggle_tables(text, tables)
    
    res = re.sub(r"\n{3,}", "\n\n", text).strip()
    _stats.data["chars"][1] += len(res)
    return res

# Handlers
def h_pdf(p: Path) -> Optional[str]:
    if not PYMUPDF_OK: return None
    doc = fitz.open(p)
    if (len("".join(p.get_text() for p in doc))/len(doc) < 500) and _MODES["ocr"] and OCR_OK:
        _stats.data["ocr"] += 1
        pages = []
        for pg in doc:
            with tempfile.NamedTemporaryFile(suffix=".png", delete=False) as f:
                pg.get_pixmap(dpi=300).save(f.name)
                pages.append(pytesseract.image_to_string(Image.open(f.name), lang='spa'))
        return "\n\n".join(pages)
    return pymupdf4llm.to_markdown(str(p))

def h_docx(p: Path) -> Optional[str]:
    if not DOCX_OK: return None
    doc = Document(p)
    res = [f"## {p.text.strip()}" if p.style.name.startswith('Hea') else p.text.strip() for p in doc.paragraphs if p.text.strip()]
    for t in doc.tables:
        for r in t.rows: res.append("| " + " | ".join(c.text.strip() for c in r.cells) + " |")
    return "\n".join(res)

def h_xl(p: Path) -> Optional[str]:
    if not PANDAS_OK: return None
    xl = pd.ExcelFile(p)
    return "\n".join([f"## {s}\n{pd.read_excel(p, s).to_markdown(index=False)}" for s in xl.sheet_names])

HANDLERS = {".pdf": h_pdf, ".docx": h_docx, ".xlsx": h_xl, ".xls": h_xl, ".txt": lambda p: p.read_text(errors='ignore')}

def convert(p: Path, out_dir: Path, verbose: bool = True) -> bool:
    try:
        if p.suffix.lower() not in HANDLERS: return False
        if verbose: print(f"📖 v6.1: {p.name}")
        raw = HANDLERS[p.suffix.lower()](p)
        if not raw: return False
        final = get_fm(p) + clean_md(raw, verbose)
        target = out_dir / normalize_fn(p.name)
        target.parent.mkdir(parents=True, exist_ok=True)
        target.write_text(final, encoding='utf-8')
        _stats.data["files"][p.suffix.lower()] = _stats.data["files"].get(p.suffix.lower(), 0) + 1
        return True
    except Exception as e:
        if _logger: _logger.error(f"{p.name}: {e}")
        return False

def run(src: Path, dst: Path, wrk: int):
    global _stats; _stats = Stats()
    files = [f for f in src.rglob("*") if f.suffix.lower() in HANDLERS]
    print(f"🚀 v6.1 Iniciando ({len(files)} docs)...")
    t0 = time.time()
    with ThreadPoolExecutor(wrk) as exe:
        for f in files: exe.submit(convert, f, dst / f.relative_to(src).parent, wrk==1)
    _stats.data["time"] = time.time() - t0
    _stats.show()

if __name__ == "__main__":
    import argparse
    parser = argparse.ArgumentParser()
    for n, d in [("--input", "pdf"), ("--output", "docs"), ("--workers", 1)]: parser.add_argument(n, default=d)
    for n in ["--ocr", "--debug", "--conservar-tablas"]: parser.add_argument(n, action="store_true")
    args = parser.parse_args()
    _MODES.update({"debug": args.debug, "ocr": args.ocr, "strict": args.conservar_tablas})
    run(Path(args.input), Path(args.output), int(args.workers))
