requirements.txt
pymupdf
pymupdf4llm
python-docx
pandas
python-pptx
Pillow
pytesseract
openpyxl
tabulate
